HBX-1906: Rename package 'org.hibernate.tool.internal.export.pojo' to 'org.hibernate.tool.internal.export.java'
