+++ orm/src/main/java/org/hibernate/tool/internal/export/java/Cfg2JavaTool.java
@@ -2,7 +2,7 @@
  * Created on 02-Dec-2004
  *
  */
-package org.hibernate.tool.internal.export.pojo;
+package org.hibernate.tool.internal.export.java;
 
 import java.util.HashMap;
 import java.util.Iterator;
