+++ orm/src/main/java/org/hibernate/tool/internal/export/java/MetaAttributeConstants.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.internal.export.pojo;
+package org.hibernate.tool.internal.export.java;
 
 public interface MetaAttributeConstants {
