+++ orm/src/main/java/org/hibernate/tool/internal/export/java/NoopImportContext.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.internal.export.pojo;
+package org.hibernate.tool.internal.export.java;
 
 public class NoopImportContext implements ImportContext {
